merge

+++ src/corewar/read_input.c
@@ -6,7 +6,7 @@
 /*   By: nalexand <nalexand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/05 18:45:02 by nalexand          #+#    #+#             */
-/*   Updated: 2019/08/16 05:13:09 by nalexand         ###   ########.fr       */
+/*   Updated: 2019/08/16 18:00:29 by nalexand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,13 +37,13 @@ static int	puterr(const char *file, const char *error)
 
 static int	validate_input(char *input, ssize_t size, char *file)
 {
-	if (size < EXEC_CODE_OFSET)
+	if (size < (ssize_t)EXEC_CODE_OFSET)
 		return (puterr(file, " is too small to be a champion"));
-	else if (size - EXEC_CODE_OFSET > CHAMP_MAX_SIZE)
+	else if (size - (ssize_t)EXEC_CODE_OFSET > (ssize_t)CHAMP_MAX_SIZE)
 		return (put_size_error(size - EXEC_CODE_OFSET, file));
 	if (GET_VAL(input, int) != COREWAR_EXEC_MAGIC)
 		return (puterr(file, " has an invalid header"));
-	if (GET_VAL(input + CODE_SIZE_OFSET, int) != size - EXEC_CODE_OFSET)
+	if (GET_VAL(input + CODE_SIZE_OFSET, int) != size - (ssize_t)EXEC_CODE_OFSET)
 		return (puterr(file, " has a code size that differ"\
 		" from what its header says"));
 	return (1);
